fix(planner): close remaining backport gaps and meet file-size policy
Completes the review action list from the fork backport audit.

Correctness and hardening:
- MilpBoundsBuilder now writes through a declared MilpColumnLayout. Bounds are
  addressed by block name, so a wrong offset cannot be expressed rather than
  merely detected. The layout validates that blocks tile the vector exactly,
  and solver offsets are derived from it, so the constraint matrices and the
  bounds assembly cannot drift apart. Completes upstream #17's intent.
- A failed coordinator cycle no longer strands a pending newer-state flag. The
  exception previously escaped the retain loop, so an event arriving during the
  failed cycle was never acted on. Retries are bounded by
  _MAX_FAILED_UPDATE_RETRIES because, unlike the fork's debounced loop, this one
  has no sleep and would otherwise hot-spin.

Dead code and documentation:
- Remove the price_actionable parameter. PlannedSlot has no such field here, so
  the guard was permanently True and unreachable.
- Delete the orphaned models/price_source.py stub. It was added by #780 but
  imported nowhere, and memories.md records ENTSO-E (#5) as out of scope.
- Document the effective battery-origin export floor and the auditable meter
  cash-flow identity in the spec; both were already implemented.
- Correct the horizon documentation: the config flow offers 12/24/36/48/72,
  while the spec still claimed 24/48/72 only.
- Fix a stale milp_optimizer docstring still describing the removed
  export<=import and import>=0 clamps.

File-size policy (30 KB AND 1000 lines):
- All production files now comply. coordinator.py 1690 -> 351 lines via three
  behaviour mixins over a shared type-only state declaration; applier.py
  1083 -> 562 via four modules; plus splits of config_flow, ev_planner,
  milp_optimizer, _constraints, forecast_tracker, and state_collector.
- Every split is a pure move. Test patch targets were repointed to the module
  where each name is now looked up.

Tests: 2670 passing (+27). New coverage for the declared layout, the coordinator
retry, portable pricing economics, and independent financial channel continuity.
