+++ custom_components/hsem/custom_sensors/applier.py
@@ -31,7 +31,6 @@ to Home Assistant.
 
 from __future__ import annotations
 
-import re
 from typing import Any
 
 from homeassistant.const import STATE_UNAVAILABLE, STATE_UNKNOWN
@@ -40,7 +39,26 @@ from custom_components.hsem.const import (
     DEFAULT_HSEM_BATTERIES_WAIT_MODE,
     DEFAULT_HSEM_EV_CHARGER_TOU_MODES,
     DEFAULT_HSEM_TOU_MODES_FORCE_CHARGE,
-    GRID_EXPORT_LIMIT_WATT,
+)
+from custom_components.hsem.custom_sensors.applier_caps import (  # noqa: F401
+    _configured_battery_device_ids,
+    _fmt_live_power_w,
+    _should_force_export_for_ev,
+    _wait_mode_self_consumption_cap_w,
+    compute_ev_discharge_cap_w,
+)
+from custom_components.hsem.custom_sensors.applier_forcible_discharge import (  # noqa: F401
+    _async_apply_forcible_discharge,
+)
+from custom_components.hsem.custom_sensors.applier_power_control import (  # noqa: F401
+    async_apply_inverter_power_control,
+)
+from custom_components.hsem.custom_sensors.applier_state_readers import (  # noqa: F401
+    _is_watt_limit,
+    _parse_power_control_pct,
+    _read_number_state,
+    _read_select_state,
+    _read_tou_periods,
 )
 from custom_components.hsem.models.hourly_recommendation import HourlyRecommendation
 from custom_components.hsem.models.live_state import LiveState
@@ -52,15 +70,10 @@ from custom_components.hsem.utils.ha_helpers import (
     async_set_select_option,
 )
 from custom_components.hsem.utils.huawei import (
-    async_set_forcible_discharge,
-    async_set_grid_export_power_pct,
-    async_set_grid_export_power_watt,
     async_set_tou_periods,
     async_stop_forcible_discharge,
-    extract_tou_periods,
 )
 from custom_components.hsem.utils.inverter_verify import (
-    ApplyResult,
     ApplyStatus,
     CycleApplySummary,
     async_write_and_verify,
@@ -73,312 +86,6 @@ from custom_components.hsem.utils.misc import (
 from custom_components.hsem.utils.recommendations import Recommendations
 from custom_components.hsem.utils.units import slot_duration_hours
 from custom_components.hsem.utils.workingmodes import WorkingModes
-
-
-def _fmt_live_power_w(power_w: float | None) -> str:
-    """Format a live power reading for log lines (``None`` → ``n/a``)."""
-    if power_w is None:
-        return "n/a"
-    return f"{int(power_w)} W"
-
-
-def _configured_battery_device_ids(cfg: SensorConfig) -> list[str]:
-    """Return configured battery device IDs, preserving order and uniqueness."""
-    device_ids: list[str] = []
-    for device_id in (
-        cfg.huawei_solar_device_id_batteries,
-        cfg.huawei_solar_device_id_batteries_2,
-    ):
-        if device_id and device_id not in device_ids:
-            device_ids.append(device_id)
-    return device_ids
-
-
-def _wait_mode_self_consumption_cap_w(
-    battery_capacity_kwh: float,
-    required_capacity_kwh: float,
-    slot_hours: float,
-    max_discharge_power_w: int,
-) -> int:
-    """Return the discharge power cap for wait-mode self-consumption.
-
-    When the battery holds more energy than the planner has reserved for future
-    expensive periods, the surplus may be used for normal household
-    self-consumption.  The cap is the power required to consume exactly that
-    surplus over the slot duration; the inverter will only draw what the house
-    actually needs, so the battery will not discharge faster than the surplus
-    allows.
-
-    Args:
-        battery_capacity_kwh: Current usable battery energy above the discharge
-            floor (kWh).
-        required_capacity_kwh: Energy the planner has reserved for future use
-            (kWh).
-        slot_hours: Duration of the current recommendation slot in hours.
-        max_discharge_power_w: Maximum discharge power supported by the battery
-            pack (W).
-
-    Returns:
-        Discharge power cap in watts.  ``0`` when there is no surplus above the
-        reserve or the slot duration is invalid.
-    """
-    surplus_kwh = max(battery_capacity_kwh - required_capacity_kwh, 0.0)
-    if surplus_kwh <= 1e-9 or slot_hours <= 1e-9:
-        return 0
-    cap_w = int(surplus_kwh / slot_hours * 1000.0)
-    return min(cap_w, max_discharge_power_w)
-
-
-def compute_ev_discharge_cap_w(
-    *,
-    live_net_w: float | None,
-    ev_power_available: bool,
-    historical_w: int,
-    sub_window_ws: list[int],
-) -> int:
-    """Compute the EV discharge cap in Watts (pure function, unit-testable).
-
-    The cap limits battery discharge to the house-only load while an EV is
-    charging, so 100 % of the EV load goes to the grid.
-
-    Selection rules:
-
-    - **Live reading available** (EV power sensor present): the historical
-      baseline is the stable reference — the cap **is** the baseline.  The
-      live reading (``house_w − ev_w``) must not move the cap in either
-      direction: downward it ratchets toward zero when the CT clamp and the
-      EV sensor disagree (beta8: 363→40 W staircase), and upward it swings
-      with ordinary house noise (cooking, heat pump cycles), slowly
-      draining the battery into what is supposed to be a grid-served EV
-      session (v6.2.0-beta1: 652→1968→928 W swings emptied the battery
-      before the 06:00 scheduled plan).  A short house spike covered from
-      the grid costs a few øre; an empty battery at 06:00 costs the whole
-      morning peak.
-    - **No live reading** (boolean-only EV sensor): fall back to the
-      smallest positive sub-window average — the 1d window recalibrates
-      fastest after an upgrade or sensor configuration change.
-    - **No history at all** (fresh install): trust the live reading.
-
-    Args:
-        live_net_w: ``net_consumption_w`` from the live state (EV power
-            already subtracted), or ``None``.
-        ev_power_available: Whether at least one EV power sensor reported
-            a positive reading this cycle.
-        historical_w: House baseline in Watts from the current slot's
-            weighted average (0 when unavailable).
-        sub_window_ws: Sub-window averages (1d/3d/7d/14d/weighted)
-            converted to Watts.
-
-    Returns:
-        The discharge cap in Watts (≥ 0).
-    """
-    if live_net_w is not None and ev_power_available:
-        if historical_w > 0:
-            return historical_w
-        return int(max(live_net_w, 0.0))
-    best_w = 0
-    for w in sub_window_ws:
-        if w > 0 and (best_w == 0 or w < best_w):
-            best_w = w
-    return best_w
-
-
-def _should_force_export_for_ev(
-    ev: Any,
-    ev_cfg: Any,
-    live: LiveState,
-) -> bool:
-    """Return True if the EV needs charging and export should be forced."""
-    if not ev.is_connected:
-        return False
-    if (
-        isinstance(ev.soc_pct, (int, float))
-        and isinstance(ev.soc_target_pct, (int, float))
-        and ev.soc_pct < ev.soc_target_pct
-    ):
-        return True
-    if (
-        isinstance(ev.soc_pct, (int, float))
-        and ev_cfg.allow_charge_past_target_soc
-        and ev.soc_pct < 100
-        and live.huawei_batteries_soc_pct is not None
-        and live.huawei_batteries_soc_pct >= 99.0
-    ):
-        return True
-    return False
-
-
-async def async_apply_inverter_power_control(
-    sensor: Any,  # NOSONAR -- HA internal type; circular import risk
-    cfg: SensorConfig,
-    live: LiveState,
-) -> CycleApplySummary:
-    """Set the grid-export power limit on all inverters.
-
-    The inverter grid connection point is controlled by price and by the
-    user-configured grid export cap:
-
-    - Negative export price → block all export with a soft watt floor
-      (``GRID_EXPORT_LIMIT_WATT``), because exporting then costs money.
-    - Non-negative export price → allow export, but cap it at
-      ``cfg.max_grid_export_power_kw`` when that value is configured.  A cap of
-      ``0`` or unset is treated as unlimited/100 %.  Battery-to-grid export
-      below ``export_electricity_min_price`` is gated planner-side by the
-      MILP and discharge scheduler, not by throttling the whole connection point.
-
-    This avoids the issue described in #767, where a positive-but-low export
-    price caused the applier to write a 100 W connection-point limit that
-    blocked surplus PV export once the battery was full.
-
-    Only issues a hardware write when the inverter state actually needs to change.
-
-    Each write is wrapped with :func:`~utils.inverter_verify.async_write_and_verify`
-    so that the inverter is polled after the write and the result is verified
-    within tolerance.  If any write fails all retries, further writes within this
-    cycle are blocked and the failure is recorded in the returned summary.
-
-    This function includes its own safety gate as defense-in-depth.  Callers
-    (``working_mode_sensor``) are expected to gate writes too, but this
-    secondary check ensures no write ever reaches the inverter when
-    ``cfg.read_only`` is ``True`` or the degraded mode is ``Error``.
-
-    Args:
-        sensor: ``HSEMWorkingModeSensor`` instance for HA access and logging.
-        cfg: Current sensor configuration.
-        live: Live state snapshot (prices, EV states, inverter control state).
-
-    Returns:
-        :class:`CycleApplySummary` with one :class:`ApplyResult` per inverter
-        write attempted.  Returns an empty summary immediately when blocked.
-    """
-    summary = CycleApplySummary()
-
-    # Defense-in-depth: block writes if read_only or degraded mode is Error.
-    if cfg.read_only:
-        _LOGGER.debug("async_apply_inverter_power_control: skipped — read_only=True")
-        return summary
-    if not hardware_writes_allowed(live.degraded_mode):
-        _LOGGER.debug(
-            "async_apply_inverter_power_control: skipped — degraded mode: %s",
-            live.degraded_mode.value,
-        )
-        return summary
-
-    export_price = live.export_electricity_price
-    min_price = cfg.export_electricity_min_price
-
-    if not isinstance(export_price, (int, float)):
-        return summary
-    if not isinstance(min_price, (int, float)):
-        return summary
-
-    # Negative export prices are the only case where we physically block the
-    # whole grid connection point.  When exporting costs money we must not
-    # allow any export, including surplus PV.  For all non-negative prices we
-    # keep the connection point open and let the planner gate battery-to-grid
-    # export via export_electricity_min_price (issue #767).
-    if export_price < 0.0:
-        desired = GRID_EXPORT_LIMIT_WATT
-        desired_is_watt = True
-        _LOGGER.debug(
-            "Export price %.4f is negative; blocking all grid export with %d W limit.",
-            export_price,
-            desired,
-        )
-    else:
-        grid_export_cap_kw = cfg.max_grid_export_power_kw
-        if grid_export_cap_kw > 1e-9:
-            # Respect the configured DNO/grid export limit as a hard cap.
-            desired = int(round(grid_export_cap_kw * 1000.0))
-            desired_is_watt = True
-            _LOGGER.debug(
-                "Export price %.4f is non-negative; allowing export up to "
-                "configured grid limit %d W (%.3f kW).",
-                export_price,
-                desired,
-                grid_export_cap_kw,
-            )
-        else:
-            # No cap configured → unlimited/100 % export.
-            desired = 100
-            desired_is_watt = False
-            if export_price < min_price:
-                _LOGGER.debug(
-                    "Export price %.4f is below export_electricity_min_price %.4f; "
-                    "leaving grid feed-in limit unlimited so PV surplus can export. "
-                    "Battery-to-grid export is gated by the planner.",
-                    export_price,
-                    min_price,
-                )
-
-    _LOGGER.debug(
-        "Determined export power limit: %s%s (export=%s, min=%s, "
-        "ev1_connected=%s, ev2_connected=%s)",
-        desired,
-        "W" if desired_is_watt else "%",
-        export_price,
-        min_price,
-        live.ev.is_connected,
-        live.ev_second.is_connected,
-    )
-
-    current_pct = _parse_power_control_pct(live.huawei_inverter_active_power_control)
-    current_is_watt = _is_watt_limit(live.huawei_inverter_active_power_control)
-
-    for inv_id in [
-        cfg.huawei_solar_device_id_inverter_1,
-        cfg.huawei_solar_device_id_inverter_2,
-    ]:
-        if inv_id is None:
-            continue
-
-        inv_entity = cfg.huawei_solar_inverter_active_power_control
-        reader_fn = lambda inv=inv_entity: _parse_power_control_pct(
-            sensor.hass.states.get(inv).state
-            if inv and sensor.hass.states.get(inv) is not None
-            else None
-        )
-
-        # Skip if the inverter already matches the desired state.
-        if (
-            current_pct is not None
-            and current_is_watt == desired_is_watt
-            and current_pct == desired
-        ):
-            continue
-
-        if desired_is_watt:
-            result = await async_write_and_verify(
-                entity_id=inv_entity or f"inverter:{inv_id}",
-                desired=desired,
-                writer=lambda _id=inv_id, _w=desired: async_set_grid_export_power_watt(  # type: ignore[misc]  # mypy cannot infer lambda types with default parameters
-                    sensor, _id, _w
-                ),
-                reader=reader_fn,
-            )
-        else:
-            result = await async_write_and_verify(
-                entity_id=inv_entity or f"inverter:{inv_id}",
-                desired=desired,
-                writer=lambda _id=inv_id, _pct=desired: (  # type: ignore[misc]  # mypy cannot infer lambda types with default parameters
-                    async_set_grid_export_power_pct(sensor, _id, _pct)
-                ),
-                reader=reader_fn,
-            )
-
-        summary.results.append(result)
-
-        if result.status == ApplyStatus.FAILED:
-            mode = "W" if desired_is_watt else "%"
-            _LOGGER.debug(
-                "Export power %s write FAILED for inverter %s after all retries. "
-                "Blocking further writes this cycle.",
-                mode,
-                inv_id,
-            )
-            return summary
-
-    return summary
 
 
 async def async_apply_battery_settings(
@@ -853,231 +560,3 @@ async def async_apply_battery_settings(
             return summary
 
     return summary
-
-
-# ---------------------------------------------------------------------------
-# Private helpers
-# ---------------------------------------------------------------------------
-
-
-async def _async_apply_forcible_discharge(
-    sensor: Any,  # NOSONAR -- HA internal type; circular import risk
-    cfg: SensorConfig,
-    live: LiveState,
-    current_required_kwh: float,
-    max_discharge_power: int,
-) -> list[ApplyResult]:
-    """Issue a forcible-discharge command to the battery pack and verify acceptance.
-
-    Returns:
-        List of :class:`ApplyResult` entries (one per configured battery device).
-        Returns an empty list if preconditions are not met and no write is attempted.
-    """
-    battery_device_ids = _configured_battery_device_ids(cfg)
-    if (
-        live.battery_usable_capacity_kwh <= 0
-        or current_required_kwh < 0
-        or not battery_device_ids
-    ):
-        return []
-
-    target_soc = int(
-        live.huawei_batteries_end_of_discharge_soc_pct
-    )  # discharge to floor
-    target_soc = max(5, min(100, target_soc))  # clamp 5-100 for safety
-
-    bat_fc_entity = cfg.huawei_solar_batteries_forcible_charge
-
-    def _read_fc_accepted() -> float | None:
-        """Return 1.0 if forcible charge state is active (not stopped/empty),
-        None otherwise.  The forcible_charge sensor reports a string like
-        'Discharging at 5000W until 5.0%' when active, or 'Stopped' when idle."""
-        if not bat_fc_entity:
-            return None
-        state = sensor.hass.states.get(bat_fc_entity)
-        if state is None or state.state in (STATE_UNKNOWN, STATE_UNAVAILABLE, ""):
-            return None
-        if state.state.lower() == "stopped":
-            return None
-        return 1.0
-
-    results: list[ApplyResult] = []
-    for device_id in battery_device_ids:
-
-        async def _write_fc(_dev: str = device_id) -> None:
-            await async_set_forcible_discharge(
-                sensor,
-                _dev,
-                target_soc,
-                max_discharge_power,
-            )
-
-        result = await async_write_and_verify(
-            entity_id=(bat_fc_entity or "forcible_charge") + f":{device_id}",
-            desired=1.0,
-            writer=_write_fc,
-            reader=_read_fc_accepted,
-            # The forcible_charge sensor changes state immediately when the
-            # command is accepted — no need for wide tolerance or retries.
-            tolerance=0.0,
-            max_retries=3,
-        )
-        results.append(result)
-        _LOGGER.debug(
-            "Excess battery export: Set forcible discharge for device %s to %d%% "
-            "SOC at %dW power. Verify result: %s",
-            device_id,
-            target_soc,
-            max_discharge_power,
-            result.status.value,
-        )
-        if result.status == ApplyStatus.FAILED:
-            break
-    return results
-
-
-# ---------------------------------------------------------------------------
-# Read-back helpers (pure — no side effects)
-# ---------------------------------------------------------------------------
-
-
-def _read_number_state(
-    sensor: Any, entity_id: str | None
-) -> float | None:  # NOSONAR -- HA internal type; circular import risk
-    """Read a number entity state from HA and return it as float, or None.
-
-    Args:
-        sensor: HSEM sensor instance with a ``hass`` attribute.
-        entity_id: HA entity ID to read.
-
-    Returns:
-        Current numeric state, or ``None`` when the entity is unavailable.
-    """
-    if not entity_id:
-        return None
-    state = sensor.hass.states.get(entity_id)
-    if state is None or state.state in (STATE_UNKNOWN, STATE_UNAVAILABLE, None):
-        return None
-    try:
-        return float(state.state)
-    except ValueError, TypeError:
-        return None
-
-
-def _read_tou_periods(
-    sensor: Any, entity_id: str | None
-) -> list[str] | None:  # NOSONAR -- HA internal type; circular import risk
-    """Read the live TOU schedule from HA and return it as a period list.
-
-    The schedule lives in the entity's ``Period 1``…``Period 10`` attributes.
-    The entity *state* is only the number of configured periods, so it can
-    never be used to verify a written schedule.  This reads HA directly rather
-    than :class:`LiveState`, whose snapshot is captured before the write and
-    would therefore never reflect it.
-
-    Args:
-        sensor: HSEM sensor instance with a ``hass`` attribute.
-        entity_id: HA entity ID to read.
-
-    Returns:
-        Current period strings, or ``None`` when the entity is unavailable.
-    """
-    if not entity_id:
-        return None
-    state = sensor.hass.states.get(entity_id)
-    if state is None or state.state in (STATE_UNKNOWN, STATE_UNAVAILABLE, None):
-        return None
-    return extract_tou_periods(state.attributes)
-
-
-def _read_select_state(
-    sensor: Any, entity_id: str | None
-) -> str | None:  # NOSONAR -- HA internal type; circular import risk
-    """Read a select entity state from HA and return it as a string, or None.
-
-    Args:
-        sensor: HSEM sensor instance with a ``hass`` attribute.
-        entity_id: HA entity ID to read.
-
-    Returns:
-        Current option string, or ``None`` when unavailable.
-    """
-    if not entity_id:
-        return None
-    state = sensor.hass.states.get(entity_id)
-    if state is None or state.state in (STATE_UNKNOWN, STATE_UNAVAILABLE, None):
-        return None
-    return str(state.state)
-
-
-def _parse_power_control_pct(state: str | None) -> int | None:
-    """Parse the inverter active power control state string into a numeric value.
-
-    Handles both percentage (``"Limited to 80%"`` → 80) and watt-based
-    (``"Limited to 100W"`` → 100) formats.  ``"Unlimited"`` returns 100.
-
-    Args:
-        state: Raw string from the inverter entity (e.g. ``"Unlimited"``,
-               ``"Limited to 80%"``, or ``"Limited to 100W"``).
-
-    Returns:
-        Integer value (percentage or watts), or ``None`` if the string
-        cannot be parsed.
-    """
-    if not isinstance(state, str):
-        return None
-    normalized = state.strip().lower()
-    # Accept any locale-independent representation of "unlimited" / no cap.
-    if normalized in (
-        "unlimited",
-        "ikke begrænset",
-        "onbeperkt",
-        "unbegrenzt",
-        "illimitato",
-        "sin límite",
-        "không giới hạn",
-    ):
-        return 100
-    # Extract the numeric value regardless of surrounding translated text or
-    # unit suffix (% or W).  This handles patterns like:
-    #   "Limited to 80%"   →  80
-    #   "Limited to 100W"  →  100
-    #   "Begrenzt auf 80 %"  →  80
-    #   "Beperkt tot 80%"  →  80
-    match = re.search(r"(-?\d+(?:\.\d+)?)", normalized)
-    if match:
-        try:
-            return int(round(float(match.group(1))))
-        except ValueError, TypeError:
-            pass
-    return None
-
-
-def _is_watt_limit(state: str | None) -> bool:
-    """Check if the power control state represents a watt-based limit.
-
-    Args:
-        state: Raw string from the inverter entity (e.g. ``"Limited to 100W"``
-               or ``"Limited to 80%"``).
-
-    Returns:
-        ``True`` if the state is a watt-based limit, ``False`` otherwise
-        (percentage-based or unlimited).
-    """
-    if not isinstance(state, str):
-        return False
-    normalized = state.strip().lower()
-    # Unlimited / percentage-based states never contain a watt indicator
-    if normalized in (
-        "unlimited",
-        "ikke begrænset",
-        "onbeperkt",
-        "unbegrenzt",
-        "illimitato",
-        "sin límite",
-        "không giới hạn",
-    ):
-        return False
-    # Look for a number immediately followed (with optional whitespace) by "w"
-    # Single quantifier avoids polynomial backtracking from stacked greedy quantifiers
-    return bool(re.search(r"\d[\d\s]*w", normalized))
