+++ custom_components/hsem/planner/milp_optimizer.py
@@ -12,7 +12,8 @@ Objective: Σ p_imp·gi - p_exp·ge + cycle_cost·m + p_soc·penalties.
 Constraints: SoC recurrence, SoC soft bounds, charge/discharge limits,
 mutex, energy balance (with efficiencies), EV co-optimisation, fuse limit.
 
-Price sanitisation: battery-export floors, export≤import, import_obj≥0.
+Price sanitisation: battery-export floors only; finite signed import and
+export rates are preserved (no export≤import or import≥0 clamp).
 Curtailment variable ``curt[t]`` allows explicit PV shedding.
 
 Pure Python, no HA imports — testable with plain pytest.
@@ -24,13 +25,14 @@ from datetime import datetime
 from typing import TYPE_CHECKING
 
 from custom_components.hsem.models.ev_config import EVConfig
+from custom_components.hsem.planner._scipy_probe import (  # noqa: F401
+    is_scipy_available,
+)
 from custom_components.hsem.utils.datetime_utils import as_tz
 from custom_components.hsem.utils.logger import log_planner
 from custom_components.hsem.utils.misc import clamp_efficiency
 from custom_components.hsem.utils.units import (
-    fuse_max_energy_per_slot_kwh,
     slot_duration_hours,
-    timedelta_to_hours,
 )
 
 if TYPE_CHECKING:
@@ -266,7 +268,17 @@ def solve_milp(
     p_imp = np.array([slots[i].price.import_price for i in future_idx], dtype=float)
     p_exp = np.array([slots[i].price.export_price for i in future_idx], dtype=float)
 
+    from custom_components.hsem.planner.milp._fuse_setup import (
+        resolve_fuse_import_limit_kwh,
+    )
+    from custom_components.hsem.planner.milp._layout import (
+        build_milp_column_layout,
+        derive_milp_offsets,
+    )
     from custom_components.hsem.planner.milp._price_sanitise import sanitize_prices
+    from custom_components.hsem.planner.milp._reserve_diagnostics import (
+        attach_export_reserve_diagnostics,
+    )
 
     p_imp_obj, p_exp, battery_export_blocked = sanitize_prices(
         p_imp,
@@ -405,50 +417,37 @@ def solve_milp(
     #   + [evN_c(0..m-1) for each active EV]      ← EV DC charge per slot
     #   + [evN_target_pen for each active EV]      ← deadline target slack
     # ------------------------------------------------------------------
-    ec_off, ed_off, gi_off, ge_off, pv_off, m_off = 0, m, 2 * m, 3 * m, 4 * m, 5 * m
-    s_max_off = 6 * m
-    s_min_off = 7 * m
-    curt_off = 8 * m
-    battery_export_off = 9 * m
-    export_mode_off = 10 * m
-    grid_flow_mode_off = 11 * m
-    n_vars = 12 * m
-
-    # --- EV variable layout ---
-    ev_var_offsets: list[int] = []  # start of ev_c[t] block per EV
-    ev_pen_offsets: list[int] = []  # index of deadline penalty per EV
-    for _ev_idx, _ev in enumerate(active_evs):
-        ev_var_offsets.append(n_vars)
-        n_vars += m  # ev_c[0..m-1] per EV
-        ev_pen_offsets.append(n_vars)
-        n_vars += 1  # single penalty per EV
+    # The declared layout is the single source of truth for the decision-vector
+    # shape; every offset below is read from it rather than recomputed by hand,
+    # so the constraint matrices and the bounds assembly cannot drift apart.
+    fuse_active = main_fuse_amps is not None and main_fuse_amps > 1e-9
+    column_layout = build_milp_column_layout(
+        m,
+        len(active_evs),
+        fuse_active=fuse_active,
+    )
+    _off = derive_milp_offsets(column_layout, len(active_evs))
+    n_vars = _off.n_vars
+    ec_off, ed_off, gi_off, ge_off = _off.ec_off, _off.ed_off, _off.gi_off, _off.ge_off
+    pv_off, m_off = _off.pv_off, _off.m_off
+    s_max_off, s_min_off, curt_off = _off.s_max_off, _off.s_min_off, _off.curt_off
+    battery_export_off = _off.battery_export_off
+    export_mode_off = _off.export_mode_off
+    grid_flow_mode_off = _off.grid_flow_mode_off
+    ev_var_offsets = _off.ev_var_offsets
+    ev_pen_offsets = _off.ev_pen_offsets
 
     # --- Fuse constraint variables ---
-    # When main_fuse_amps is provided and > 0, add gi_pen[t] penalty
-    # variables that absorb grid import exceeding the fuse rating.
-    fuse_active = main_fuse_amps is not None and main_fuse_amps > 1e-9
+    # When main_fuse_amps is provided and > 0, gi_pen[t] penalty variables
+    # absorb grid import exceeding the fuse rating.
     if fuse_active:
-        gi_pen_off = n_vars
-        n_vars += m  # gi_pen[0..m-1] per slot
-        # Calculate max grid import per slot in kWh (single source of truth
-        # shared with the post-hoc EV/battery throttle in engine_core).
-        # We derive interval_minutes from the first slot's duration.
+        gi_pen_off = column_layout.offset("grid_import_penalty")
         first_slot = slots[future_idx[0]]
-        interval_minutes = timedelta_to_hours(first_slot.end - first_slot.start) * 60.0
-        assert main_fuse_amps is not None  # guarded by fuse_active
-        max_grid_import_per_slot_kwh = fuse_max_energy_per_slot_kwh(
-            main_fuse_amps,
-            main_fuse_phases,
-            interval_minutes / 60.0,
-        )
-        log_planner(
-            "debug",
-            "[milp] Main fuse constraint active: %d A × %d-phase → max %.3f kWh/slot "
-            "(interval=%.0f min)",
-            main_fuse_amps,
-            main_fuse_phases,
-            max_grid_import_per_slot_kwh,
-            interval_minutes,
+        max_grid_import_per_slot_kwh = resolve_fuse_import_limit_kwh(
+            main_fuse_amps=main_fuse_amps,
+            main_fuse_phases=main_fuse_phases,
+            slot_start=first_slot.start,
+            slot_end=first_slot.end,
         )
     else:
         gi_pen_off = 0  # unused when fuse is inactive
@@ -555,6 +554,7 @@ def solve_milp(
         _has_session_demand,
         session_slot_hours=session_slot_hours,
         available_slot_hours=available_slot_hours,
+        column_layout=column_layout,
         max_grid_export_per_slot_kwh=max_grid_export_per_slot_kwh,
         export_limit_active=export_limit_active,
         battery_export_blocked=battery_export_blocked,
@@ -734,46 +734,15 @@ def solve_milp(
     diagnostics["primary_postwrite_inventory_validation"] = inventory_validation
     if ev_writeback_diagnostics:
         diagnostics["ev"] = ev_writeback_diagnostics
-    diagnostics["battery_export_reserve_active"] = bool(
-        constraints.get("battery_export_reserve_active", False)
+    attach_export_reserve_diagnostics(
+        diagnostics,
+        constraints,
+        m=m,
+        export_mode_off=export_mode_off,
+        solution=result.x,
+        ec_sol=ec_sol,
+        ed_sol=ed_sol,
+        current_kwh=current_kwh,
     )
-    checkpoints = constraints.get("export_reserve_checkpoints")
-    if diagnostics["battery_export_reserve_active"] and checkpoints is not None:
-        active_soc: list[float] = []
-        for t in range(m):
-            if result.x[export_mode_off + t] < 0.5:
-                continue
-            checkpoint = int(checkpoints[t])
-            checkpoint_soc = current_kwh + float(
-                np.sum(ec_sol[: checkpoint + 1]) - np.sum(ed_sol[: checkpoint + 1])
-            )
-            active_soc.append(checkpoint_soc)
-        diagnostics["battery_export_reserve_min_checkpoint_soc_kwh"] = (
-            round(min(active_soc), 6) if active_soc else None
-        )
 
     return out_slots, diagnostics
-
-
-def is_scipy_available() -> bool:
-    """Return ``True`` if scipy is importable in the current environment.
-
-    The import result is cached at module level so that the blocking
-    ``import scipy.optimize`` happens exactly once at import time rather
-    than on every planner run inside the Home Assistant event loop.
-    """
-    return _SCIPY_AVAILABLE
-
-
-# --- Module-level cache: computed once at import time --------------------
-def _check_scipy() -> bool:
-    """Check whether scipy is importable.  Called once at module load."""
-    try:
-        import scipy.optimize  # noqa: F401
-
-        return True
-    except ImportError:
-        return False
-
-
-_SCIPY_AVAILABLE: bool = _check_scipy()
