+++ tests/test_ha_mock_integration.py
@@ -2078,12 +2078,12 @@ def _patch_all_ha_helpers():
                 return_value=None,
             ),
             patch(
-                "custom_components.hsem.coordinator"
+                "custom_components.hsem.coordinator_cycle"
                 ".populate_avg_house_consumption_from_snapshot",
                 return_value=True,
             ),
             patch(
-                "custom_components.hsem.coordinator"
+                "custom_components.hsem.coordinator_cycle"
                 ".populate_price_and_solcast_from_snapshot",
             ),
             patch(
